refactor json coding functions
