+++ InDoor-User/Utils/Constants.swift
@@ -95,21 +95,3 @@ class Constants {
     static let delete = "Delete"
     static let edit = "Edit"
 }
-
-
-func encodeToJson(objectClass: Response) -> [String: Any]?{
-    do{
-        let jsonData = try JSONEncoder().encode(objectClass)
-        let json = String(data: jsonData, encoding: String.Encoding.utf8)!
-        return jsonToDictionary(from: json)
-    }catch let error{
-        print(error.localizedDescription)
-        return nil
-    }
-}
-
-func jsonToDictionary(from text: String) -> [String: Any]? {
-    guard let data = text.data(using: .utf8) else { return nil }
-    let anyResult = try? JSONSerialization.jsonObject(with: data, options: [])
-    return anyResult as? [String : Any]
-}
