+++ InDoor-User/Utils/jsonCoding.swift
@@ -1,0 +1,25 @@
+//
+//  jsonCoding.swift
+//  InDoor-User
+//
+//  Created by Mac on 15/06/2023.
+//
+
+import Foundation
+
+func encodeToJson(objectClass: Response) -> [String: Any]?{
+    do{
+        let jsonData = try JSONEncoder().encode(objectClass)
+        let json = String(data: jsonData, encoding: String.Encoding.utf8)!
+        return jsonToDictionary(from: json)
+    }catch let error{
+        print(error.localizedDescription)
+        return nil
+    }
+}
+
+func jsonToDictionary(from text: String) -> [String: Any]? {
+    guard let data = text.data(using: .utf8) else { return nil }
+    let anyResult = try? JSONSerialization.jsonObject(with: data, options: [])
+    return anyResult as? [String : Any]
+}
